+++ src/Biogeochemistry/Advection/Advection.jl
@@ -8,7 +8,8 @@ using KernelAbstractions
 
 using PlanktonIndividuals.Grids
 using PlanktonIndividuals.Architectures: device, array_type, Architecture
-using PlanktonIndividuals.Fields
+using PlanktonIndividuals.Biogeochemistry: tracer_names
+using PlanktonIndividuals.Biogeochemistry: fill_halo_Gcs!, fill_halo_tracer!
 
 include("operaters.jl")
 include("tracer_diffusion.jl")
