+++ src/Biogeochemistry/Biogeochemistry.jl
@@ -3,12 +3,25 @@ module Biogeochemistry
 export generate_tracers
 export tracers_init, default_tracer_init
 export tracer_update!
+export Field
+export interior, zero_fields!
+export fill_halo_vel!
+export default_bcs, getbc
+export set_bc!, validate_bcs
+export tracer_names
 
 using KernelAbstractions
 
 using PlanktonIndividuals.Architectures: device, array_type, Architecture
 using PlanktonIndividuals.Grids
-using PlanktonIndividuals.Fields
+
+using PlanktonIndividuals: BoundaryConditions
+
+const tracer_names=(:DIC,:NH4,:NO3,:PO4,:FeT,:DOC,:DON,:DOP,:DOFe,:POC,:PON,:POP,:POFe)
+
+include("HaloRegions/halo_regions.jl")
+include("HaloRegions/boundary_conditions.jl")
+include("HaloRegions/apply_bcs.jl")
 
 include("Advection/Advection.jl")
