+++ src/Diagnostics/Diagnostics.jl
@@ -8,7 +8,6 @@ using KernelAbstractions
 
 using PlanktonIndividuals.Architectures: device, Architecture, array_type
 using PlanktonIndividuals.Grids
-using PlanktonIndividuals.Fields
 
 using PlanktonIndividuals: AbstractMode, CarbonMode, QuotaMode, MacroMolecularMode, IronEnergyMode
 using PlanktonIndividuals: individuals, phytoplankton, abiotic_particle
