+++ src/Output/Output.jl
@@ -8,9 +8,9 @@ using LinearAlgebra: dot
 using Statistics, JLD2, Serialization, Printf
 
 using PlanktonIndividuals.Grids
-using PlanktonIndividuals.Fields
 using PlanktonIndividuals.Diagnostics
 using PlanktonIndividuals.Model
+using PlanktonIndividuals.Biogeochemistry
 
 using PlanktonIndividuals: AbstractMode, CarbonMode, QuotaMode, MacroMolecularMode, IronEnergyMode
 using PlanktonIndividuals: individuals, phytoplankton, abiotic_particle
